Add plot funcs to jumeg_plot. Minor changes.
